requirements.txt
spotipy
# gmusicapi
bumpversion
pymaybe
fuzzywuzzy
python-Levenshtein
kafka-python